refactor: Switch from filtering to walking "dirty" files
Filtering content pages is the wrong approach - it is opt-out of rendering. We need opt-in to rendering. That is why files are now marked as dirty if any `DirtyChecker` considers it dirty, and consequently that file is rendered.

The file walker now yields `FilePath` (in order to process them only once) for the dirty files. So no more bailing out of the `Processor` with an `Error`.

There is now a `getPage` function for getting a content page. The `Processor` now uses that. This new function will be used when getting pages for components and layouts.