fix: Filter current page from `wantedPages` array
The expectation is reasonably that a page wants other pages than itself, irrespective of the provided glob. For instance a folder of blog posts will include a `blog/index.md` file, but getting a list of blog posts with the `blog/*.md` glob should not include this index file.

This behavior is slightly magic, and might change in the future.