refactor: Create generic `Content` getter in `api.ts`
The main workflow file now includes a function to get generic `Content` from a `FilePath`. This function is utilized in the workflow as well as the pages getter intended for layout files. Content can now be either regular pages or static content. Note that the actual functionality is not yet changed, as all files are still marked as being pages.

Content and files now have a `type` property indicating type of object.