fix: Reload layouts when running in server mode
In order to reload modules (i.e. layouts) when in server mode, a separate process is now spawned to run the build. This process uses the `bob` executable, which means bob needs to be installed in order for this to work. The re-build uses the same command line arguments as the parent process, minus the `server` argument, of course.

Spawning a process is a separate permission in Deno, `--allow-run`. This needs to be allowed on the installed bob instance.

The reason for this is that ES modules are (by specification) cached for the duration of the process. In order to really reload everything, a new process needs to be spawned. Another possibility would have been to add a search- or hash-string to the module loading path, but that has two problems: i) it's pretty cumbersome to set up correctly, and ii) it doesn't work for changes in modules down the dependency graph (i.e. only works for the actually loaded module, which is the layout).

When Deno implements hooks for module resolution, this spawning can be avoided.