wip: Create more modular proxy for content dependencies
Now there's a wrapper for `PagesGetter` that will require almost no changes in existing code. Just use the wrapper as a page getter and commit the results after processing the request. This approach is also much more testable, and a trivial test case has already been implemented, and it's passing for the current implementation.

Create a new type `Cache`, which is used by the dependency wrapper to save and retrieve dependants of a page.