wip: Implement concurrency in dependency writer
Dependency writer now uses the new `Cache.transaction` method to create a "transaction" on the cache, where the current value can be read, amended, and written while keeping a lock on the specific cache key. This means that dependencies can be updated for a key in parallel, i.e. pages can be built concurrently.

Update the main API to use the new dependency writer. Create a memory cache that implements the new `Cache` interface.