docs: Fix edge component worker build error
The Deno bundler doesn't seem to correctly tree-shake when bundling with dependencies included via `mod.ts`. Everything (including static site generation) seems to be included in the bundle. This resulted in an error, since the `Deno` global is not defined in a worker.

Use exports from `edge/mod.ts` directly instead both in worker and in the edge components themselves.

Other:
- Import domain in `edge/mod.ts` to get intrinsic elements in scope for edge components
- Ignore edge component docs `dist` directory, which is where the Worker bundle is cached
- Add `Record<string, unknown>` to `EdgeElement` props to avoid type warning