feat: Add server functions beta
Functions are now loaded dynamically from `functions/index.ts` when the functions server is loaded. Function path patterns are matched against the request URL with `URLPattern` and the first found function is executed and the result returned to the client. If no function is found, 404 is returned.

This commit includes the following changes:

- Move `loadIfExists` to separate file in order to use the same logic when loading functions
- Add `bob functions` for running functions server
- Start functions on `bob server`
- Remove the build shell script from the documentation site, as it's unnecessary
- Add test functions for the docs site
- Add preliminary documentation for the functions feature