Get dependants in changes applier
This makes it much easier to work with changes and rendering. The renderer only needs to know how to render, and a change represents a file (content) modification.

The main driver for the change was that without it, e.g. an index page with many dependencies was rendered many times.