fix: Create change as `create` when writing file from function
Sometimes you just update a file from the function, but many times a write might be a new file creation. Since arithmetically, `create` > `modify`, it should be safe to make the change as `create` instead of the previous `modify`. This results in an extra dependency check for files that are really just modified.

Closes #49